+++ src/components/BlockList/index.js
@@ -11,7 +11,8 @@ import Item from './Item';
 const BlockList = (props) => (
   <div
     className="background--white overflow-x--auto border--all"
-    data-test-section={ props.testSection }>
+    data-test-section={ props.testSection }
+    style={ { maxHeight: props.maxHeight } }>
     <ul style={ { display: 'inline-block', minWidth: '100%' } }>
       { props.children }
     </ul>
@@ -21,6 +22,14 @@ const BlockList = (props) => (
 BlockList.propTypes = {
   /** Should be subcomponents of `BlockList` */
   children: React.PropTypes.node.isRequired,
+  /**
+   * The max height of the `BlockList`. Pixels will be assumed if no unit is
+   * provided.
+   */
+  maxHeight: React.PropTypes.oneOfType([
+    React.PropTypes.number,
+    React.PropTypes.string,
+  ]),
   /** Hook for automated JavaScript tests */
   testSection: React.PropTypes.string,
 };